variable needed to be long, not int

+++ objectCode/stack64.c
@@ -7,13 +7,13 @@ long fn(long a, long b, long c, long d, long e, long f, long g)
 
 int main()
 {
-	int r =	fn(0xAAAAAAAAAAAAAAAA,
-			   0xBBBBBBBBBBBBBBBB,
-			   0xCCCCCCCCCCCCCCCC,
-			   0xDDDDDDDDDDDDDDDD,
-			   0xEEEEEEEEEEEEEEEE,
-			   0xFFFFFFFFFFFFFFFF,
-			   0x1111111111111111);
+	long r =	fn(0xAAAAAAAAAAAAAAAA,
+				   0xBBBBBBBBBBBBBBBB,
+				   0xCCCCCCCCCCCCCCCC,
+				   0xDDDDDDDDDDDDDDDD,
+				   0xEEEEEEEEEEEEEEEE,
+				   0xFFFFFFFFFFFFFFFF,
+			 	   0x1111111111111111);
 	
 	return 0;
 }
